Converted main project to netstandard2.0 and the test and console example projects to netcore3.1.
