+++ .idea/.idea.TagCloud.Core/.idea/.name
@@ -1,0 +1,1 @@
+TagCloud.Core
