+++ .idea/.idea.TagCloud.Core/.idea/modules.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/.idea/.idea.TagCloud.Core/.idea/.idea.TagCloud.Core.iml" filepath="$PROJECT_DIR$/.idea/.idea.TagCloud.Core/.idea/.idea.TagCloud.Core.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/.idea.TagCloud.Core/riderModule.iml" filepath="$PROJECT_DIR$/.idea/.idea.TagCloud.Core/riderModule.iml" />
+    </modules>
+  </component>
+</project>
