+++ Examples/ConsoleExample.Core/Sample_Text_1.txt
