+++ TagCloud.Core/TagCloudAnalyzer.cs
