+++ TagCloud.Core/TagCloudEx.cs
