--- TagCloud.Tests/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NUnit" version="3.12.0" targetFramework="net40" />
-  <package id="Should" version="1.1.20" targetFramework="net40" />
-</packages>
