--- TagCloud.sln
@@ -1,74 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29926.136
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagCloud", "TagCloud\TagCloud.csproj", "{78939909-DFC3-463F-9B5D-515C57D93BEB}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagCloud.Tests", "TagCloud.Tests\TagCloud.Tests.csproj", "{51D2C5FA-25AB-482A-89D8-CB097A26B2B0}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleExample", "Examples\ConsoleExample\ConsoleExample.csproj", "{37CC0F63-E89C-4301-9B75-EFC94CAA9C73}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebExample", "Examples\WebExample\WebExample.csproj", "{49F2E6EB-9510-4078-8356-A963CD532FFE}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|Mixed Platforms = Debug|Mixed Platforms
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|Mixed Platforms = Release|Mixed Platforms
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{78939909-DFC3-463F-9B5D-515C57D93BEB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{78939909-DFC3-463F-9B5D-515C57D93BEB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{78939909-DFC3-463F-9B5D-515C57D93BEB}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{78939909-DFC3-463F-9B5D-515C57D93BEB}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{78939909-DFC3-463F-9B5D-515C57D93BEB}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{78939909-DFC3-463F-9B5D-515C57D93BEB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{78939909-DFC3-463F-9B5D-515C57D93BEB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{78939909-DFC3-463F-9B5D-515C57D93BEB}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{78939909-DFC3-463F-9B5D-515C57D93BEB}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{78939909-DFC3-463F-9B5D-515C57D93BEB}.Release|x86.ActiveCfg = Release|Any CPU
-		{51D2C5FA-25AB-482A-89D8-CB097A26B2B0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{51D2C5FA-25AB-482A-89D8-CB097A26B2B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{51D2C5FA-25AB-482A-89D8-CB097A26B2B0}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{51D2C5FA-25AB-482A-89D8-CB097A26B2B0}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{51D2C5FA-25AB-482A-89D8-CB097A26B2B0}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{51D2C5FA-25AB-482A-89D8-CB097A26B2B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{51D2C5FA-25AB-482A-89D8-CB097A26B2B0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{51D2C5FA-25AB-482A-89D8-CB097A26B2B0}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{51D2C5FA-25AB-482A-89D8-CB097A26B2B0}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{51D2C5FA-25AB-482A-89D8-CB097A26B2B0}.Release|x86.ActiveCfg = Release|Any CPU
-		{37CC0F63-E89C-4301-9B75-EFC94CAA9C73}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{37CC0F63-E89C-4301-9B75-EFC94CAA9C73}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
-		{37CC0F63-E89C-4301-9B75-EFC94CAA9C73}.Debug|Mixed Platforms.Build.0 = Debug|x86
-		{37CC0F63-E89C-4301-9B75-EFC94CAA9C73}.Debug|x86.ActiveCfg = Debug|x86
-		{37CC0F63-E89C-4301-9B75-EFC94CAA9C73}.Debug|x86.Build.0 = Debug|x86
-		{37CC0F63-E89C-4301-9B75-EFC94CAA9C73}.Release|Any CPU.ActiveCfg = Release|x86
-		{37CC0F63-E89C-4301-9B75-EFC94CAA9C73}.Release|Mixed Platforms.ActiveCfg = Release|x86
-		{37CC0F63-E89C-4301-9B75-EFC94CAA9C73}.Release|Mixed Platforms.Build.0 = Release|x86
-		{37CC0F63-E89C-4301-9B75-EFC94CAA9C73}.Release|x86.ActiveCfg = Release|x86
-		{37CC0F63-E89C-4301-9B75-EFC94CAA9C73}.Release|x86.Build.0 = Release|x86
-		{49F2E6EB-9510-4078-8356-A963CD532FFE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{49F2E6EB-9510-4078-8356-A963CD532FFE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{49F2E6EB-9510-4078-8356-A963CD532FFE}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{49F2E6EB-9510-4078-8356-A963CD532FFE}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{49F2E6EB-9510-4078-8356-A963CD532FFE}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{49F2E6EB-9510-4078-8356-A963CD532FFE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{49F2E6EB-9510-4078-8356-A963CD532FFE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{49F2E6EB-9510-4078-8356-A963CD532FFE}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{49F2E6EB-9510-4078-8356-A963CD532FFE}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{49F2E6EB-9510-4078-8356-A963CD532FFE}.Release|x86.ActiveCfg = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1EF3E4ED-A1A7-4E50-A9F9-88E72B129EFB}
-	EndGlobalSection
-	GlobalSection(TestCaseManagementSettings) = postSolution
-		CategoryFile = TagCloud.Tests.vsmdi
-	EndGlobalSection
-EndGlobal
